code spited as mentioned in comments
